finish AdjudicateRegions.py (#19)
* clean up comments

* Add prior subfam counts file
add new input file
python3 AdjudicateRegions.py alignfile submatrix SubfamPriorCounts.txt
new data structure is dict that maps subfam name to number of times it appears in genome - will be added to confidence calculations

* Fix tests and add updated prior counts file
add subfams as input into confidence functions so can usesubfam names to index SubfamCounts dictionary
Update prior counts file to deal with L1_*
Fix failing tests because inputs to all confidence functions changed - tests should pass after confidence calculations are changed because gave them all same prior count

* add ability to input skip state align score at command line

* remove CharPos dictionary
CharPos was used to map characters in the input substitution matrix to positions in SubMatrix - instead SubMatrix is now a dict that maps char1char2 to the score. Ex - SubMatrix['AA'] = 8

* Fix bug where it was stitching subfams even when not contiguous on the consensus
In FillProbMatrix() when probability came from the same subfam but it wasn't contigous on consensus it is given a change probability, but the origin matrix says it comes from the same subfam. In GetPath() this was seeing it as all one big subfamily.
ex:
AluJr 1 600 - this is 2 Alus next to eachother
Fix: add a new matrix SameSubfamChangeMatrix that holds whether or not a position came from the same subfam, but is actually a change. Checks this in GetPath()
ex:
AluJr 1 300
AluJr 1 300

* Fix doctest in GetPath()

* Bug Fix - no edges between skip nodes
previously could add alternative edges involving skip nodes - fixed it so skip nodes will never have alternative edges.

* prior counts file now taken as optional command line argument

* update one of the test examples

* add prior counts to confidence calculations - tests will fail

* all confidence functions tests pass

* switch AlignMatrix to hold floats, also fix all type errors

* fix doc test for FillAlignScoreMatrix

* change skipAlignScore to correlate with lambda (sub matrix scaling

* support matrix now has score in the middle of region, same as align matrix

* fix FillSupportMatrix tests so they pass

* a few bug fixed from looking at examples. also added some function timing

* Fix FillPathGraph()
Was doing unecessary looping through all the subfams in a node, when only care about the sink subfams confidence in the source node. Now just index into dict directly to the subfam of interest.
Was only checking the sink subfams confidence in the source node, also added in check for the source subfams confidence in the sink node - if above a certain threshold add edge.

* change skipAlignScore to 0.0, won't be 0 in conf

* make tests pass

* Bug fix - was reporting original subfam after stitching, not updated one

* bug - same bug, new fix

* prior counts option now correct - use SubfamPriorCounts_jun11.txt for example regions

* speed up FillColumns

* make CollapseMatrices faster - was unecessarily checking in support and conensus matrices but they are the same

* make FillProbMatrix faster by creating vars to index once each

* clean up GetPath

* added function and command line option to output format for SODA visualization

* add chrom number to input alignment for SODA format

* fix tests

* fix tests

* bug fix in PrintResultsViz()

* updated naming convention in input files for soda viz output

* add soda output files for all examples

* fix FillAlignScoreMatrix()
no longer calls CalcScore() for all chunks at start of seq. now calls it once and accumulates score

* can have 0s in align matrix

* update example input files with flank info

* update soda output files for all examples

* make tests pass with changes

* reformat

* reformat

* FillConsensusMatrix speed up by only looping though cols that have seq and ignoring '.'

* make tests pass

* FillConfidenceMatrix speed up
instead of using 0s as place holders and sending all cols into confidenceCM(), use book keeping array and only send in cols with an align score

* FillSupportMatrix speed up by only looping cols where seqs are

* make tests pass

* speed up by adding ActiveCells
in FillColumns() created a dict that holds the column number and which rows are active in that column. When looping column major never have to loop through a cell that doesn't have a value in the uncollapsed matrices. Already have something similar for the collapsed matrices

* Speed up FillNodeConfidence()
instead of grabbing splices from each alignment and calculating alignments scores for all subfams in a node, just use the AlignMatrix and sum all scores/chunksize for a subfam in the node

* run esl_scorematrix if lambda not included at command line, can specify own path to easel

* add option to print support mat for heatmap and fix formatting

* add option to print results in terms of genomic position