Refactor fill_support_matrix
This refactor does two things. First, it ought to speed the function up a bit.
Second, and more importantly, it corrects a subtle bug whereby the first column
was assumed to contain data when, in reality, it does not because we want to
force the algorithm into the skip state initially.

Note, this may not (yet) handle TRs correctly. It also changes assumptions
about its inputs, so call sites will need to be updated before these changes
make it into master.