Improve debugging output
This commit refactors the printers into a class in order to hold certain pieces
of state and simplify calls to the print functions themselves.

It also improves the timeit decorator to add output options and better
messages.

Finally, it makes the ID column in the output optional (and omitted by default)
so that benchmarking and functional test outputs can be compared against
previous outputs.